+++ Editor/Exceptions.cs
@@ -6,7 +6,14 @@ namespace VRCAddonInstaller {
     public class FailedToInsertGameObject : System.Exception {
         public FailedToInsertGameObject(string message) : base(message) {
         }
-        public string pathToFbx;
-        public string pathToParent;
+        public string pathToAsset;
+        public string pathToExistingGameObject;
+    }
+
+    public class FailedToCopyBoneIntoArmature : System.Exception {
+        public FailedToCopyBoneIntoArmature(string message) : base(message) {
+        }
+        public string pathToBone;
+        public string pathToTarget;
     }
 }