+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.1
+
+- import all 3D model file formats
+- output errors
+
 # alpha2
 
 - fix Newtonsoft error